requirements_submission.txt
absl-py==2.0.0
accelerate==0.21.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
ansi2html==1.8.0
anyio==3.7.1
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
bitsandbytes==0.41.0
blinker==1.7.0
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
cachetools==5.3.2
certifi @ file:///croot/certifi_1690232220950/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
comm==0.2.0
ConfigArgParse==1.7
contourpy==1.2.0
cryptography @ file:///croot/cryptography_1694444244250/work
cycler==0.12.1
Cython==3.0.5
dash==2.14.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
deepspeed==0.9.5
diffusers==0.25.0
docker-pycreds==0.4.0
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.1.3
executing==2.0.1
fastapi==0.104.1
fastjsonschema==2.19.0
ffmpy==0.3.1
filelock @ file:///croot/filelock_1672387128942/work
flash-attn==2.3.4
Flask==3.0.0
fonttools==4.44.3
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
google-auth==2.23.4
google-auth-oauthlib==1.1.0
gradio==3.35.2
gradio_client==0.2.9
grpcio==1.59.3
h11==0.14.0
h5py==3.10.0
hjson==3.1.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.19.4
idna @ file:///croot/idna_1666125576474/work
imageio==2.33.1
importlib-metadata==6.8.0
ipython==8.17.2
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2 @ file:///croot/jinja2_1666908132255/work
joblib==1.3.2
json-tricks==3.17.3
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter_core==5.5.0
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
lazy_loader==0.3
lightning-utilities==0.9.0
linkify-it-py==2.0.2
-e git+https://github.com/haotian-liu/LLaVA.git@80540fb4bf4dad118d87d42bd2fb55e6f3b96f16#egg=llava
lmdb==1.4.1
Markdown==3.5.1
markdown-it-py==2.2.0
markdown2==2.4.10
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
msgpack==1.0.7
msgpack-numpy==0.4.7.1
multidict==6.0.4
nbformat==5.7.0
nest-asyncio==1.5.8
networkx @ file:///croot/networkx_1690561992265/work
ninja==1.11.1.1
numpy @ file:///croot/numpy_and_numpy_base_1695830428084/work/dist/numpy-1.26.0-cp310-cp310-linux_x86_64.whl#sha256=fc2732718bc9e06a7b702492cb4f5afffe9671083930452d894377bf563464a3
oauthlib==3.2.2
open3d==0.17.0
opencv-python==4.9.0.80
orjson==3.9.10
packaging==23.2
pandas==2.1.3
parso==0.8.3
peft==0.4.0
pexpect==4.8.0
Pillow @ file:///croot/pillow_1696580024257/work
platformdirs==4.0.0
plotly==5.18.0
prompt-toolkit==3.0.41
protobuf==4.23.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.13
pydub==0.25.1
Pygments==2.16.1
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.1.1
pyquaternion==0.9.9
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-lightning==2.1.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
regex==2023.10.3
requests @ file:///croot/requests_1690400202158/work
requests-oauthlib==1.3.1
retrying==1.3.4
rpds-py==0.13.0
rsa==4.9
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.2.2
scipy==1.11.3
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.35.0
setproctitle==1.3.3
shortuuid==1.0.11
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
stack-data==0.6.3
starlette==0.27.0
svgwrite==1.4.3
sympy @ file:///croot/sympy_1668202399572/work
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
threadpoolctl==3.2.0
tifffile==2023.12.9
timm==0.6.13
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchinfo==1.8.0
torchmetrics==1.2.0
torchvision==0.15.2
tqdm==4.66.1
traitlets==5.13.0
transformers==4.31.0
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
uc-micro-py==1.0.2
urllib3 @ file:///croot/urllib3_1698257533958/work
uvicorn==0.24.0.post1
wandb==0.16.0
wavedrom==2.0.3.post3
wcwidth==0.2.10
websockets==12.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
yarl==1.9.2
zipp==3.17.0